Create Server Release
The server requires a separate document as the processes is different from the client. For example binary build products are not required on Windows, Mac, Linux and Android and the Alpha testers project may not be used in the same way. There is a great deal of overlap (this is a copy and paste) so we may wish to combine in some way to meet the DRY principle.